Merged-Credet-Debiter-

+++ MAMS/MAMS_Models/Model/Purchase.cs
@@ -17,7 +17,7 @@ namespace MAMS_Models.Model
         public Decimal WeightInkg { get; set; }
         public Decimal TotalCropWeight { get; set; }
         public int PriceInMaun { get; set; }
-        public Decimal PriceInKg { get; set; }
+        public int PriceInKg { get; set; }
         public int TotalCropPrice { get; set; }
         public int TotalExp { get; set; } 
         public int TotalAmountwithExp { get; set; }
